+++ src/main/java/net/paladins/fabric/FabricMod.java
@@ -1,7 +1,6 @@
 package net.paladins.fabric;
 
 import net.fabricmc.api.ModInitializer;
-import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
 import net.fabricmc.fabric.api.loot.v2.LootTableEvents;
 import net.minecraft.item.ItemStack;
@@ -11,7 +10,6 @@ import net.paladins.item.Group;
 import net.paladins.item.PaladinItems;
 import net.paladins.item.armor.Armors;
 import net.paladins.util.SoundHelper;
-import net.paladins.worldgen.PaladinWorldGen;
 import net.spell_engine.api.loot.LootHelper;
 
 public class FabricMod implements ModInitializer {
@@ -31,7 +29,6 @@ public class FabricMod implements ModInitializer {
     }
 
     private void subscribeEvents() {
-        ServerLifecycleEvents.SERVER_STARTING.register(PaladinWorldGen::init);
         LootTableEvents.MODIFY.register((resourceManager, lootManager, id, tableBuilder, source) -> {
             LootHelper.configure(id, tableBuilder, PaladinsMod.lootConfig.value, PaladinItems.entries);
         });