+++ src/main/java/net/paladins/mixin/LivingEntityDivineProtection.java
@@ -1,4 +1,4 @@
-package net.paladins.mixin.effect;
+package net.paladins.mixin;
 
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;